backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
anthropic>=0.40.0
sqlalchemy>=2.0.35
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
pydantic>=2.10.0
psycopg2-binary>=2.9.9
